docs: add German translation to README.md
Adds German translation to README

Adds a German translation of the README file to provide documentation for German-speaking users. This allows a wider audience to understand and use the application.

Adds German translation to README

Adds a German translation of the README file.

This provides documentation for German-speaking users, broadening the application's accessibility and user base.